Moved PriceViewModelTests to PrimitivesComponents
